remove warning

+++ learningCPP/recision/uniqueSubset.cpp
@@ -7,7 +7,7 @@ class Solution
 private:
     void findSubsets(int ind, vector<int> &nums, vector<int> &ds, vector<vector<int>>& ans) {
         ans.push_back(ds);
-        for(int i = ind; i < nums.size(); i++) {
+        for(int i = ind; i < static_cast<int>(nums.size()); i++) {
             if(i!=ind && nums[i] == nums[i - 1]) continue;
             ds.push_back(nums[i]);
             findSubsets(i+1, nums, ds, ans);
